Add a balance check for newly added Whitelist address
If an address that already holds some Pandora tokens and NFTs. And suddenly, it has been set to whitelist, the result will be: its NFT will not be burned with the of Pandora tokens transfer out according to the whitelist logic. 

This bug will create additional untransferable NFTs inside Pandora's ecosystem.